Handle mystery egg in final assembly step
Moving mystery egg (and any other potential future secret
changes) to the last assembly phase will allow the bank.asm
files to be 100% readable and integrated for contributions
and pull requests. Whenever a new hack is added, a dev
with permissions must ensure the private asm is updated
as needed to resolve any conflicts.